auth-service/requirements.txt
fastapi==0.109.1
uvicorn[standard]
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
pydantic==2.7.1
